+++ src/main/java/graphics/devMode/common/DevPage.java
@@ -6,20 +6,14 @@ import java.awt.event.ActionListener;
 import org.jetbrains.annotations.Nullable;
 
 import graphicsEngine.colors.SimpleColorScheme;
-import graphicsEngine.panels.BorderProperties;
-import graphicsEngine.presets.panels.AbstractHeader;
 
 import graphics.common.CommonWindow;
 import graphics.common.CommonPage;
-import graphics.devMode.common.header.CommonHeader;
-import graphics.devMode.common.header.HeaderButtonListener;
 
 /**
  * TODO: finish this and add javadocs
  */
 public abstract class DevPage extends CommonPage {
-    private HeaderButtonListener headerButtonListener;
-
     /**
      * TODO: finish this javadoc
      */
@@ -27,24 +21,5 @@ public abstract class DevPage extends CommonPage {
                    @Nullable List<ActionListener> actionListenerList,
                    @Nullable SimpleColorScheme colors) {
         super(window, actionListenerList, colors);
-    }
-
-    //TODO: add javadoc
-    @Override
-    public boolean addParticularListener(@Nullable ActionListener listener) {
-        if (listener instanceof HeaderButtonListener) {
-            headerButtonListener = (HeaderButtonListener) listener;
-            return true;
-        }
-        return false;
-    }
-
-    /**
-     * TODO: finish this javadoc
-     */
-    @Override
-    public @Nullable AbstractHeader getHeader(@Nullable SimpleColorScheme headerColors,
-                                              @Nullable BorderProperties borderProperties) {
-        return new CommonHeader(headerColors, borderProperties, headerButtonListener);
     }
 }
