+++ src/main/java/graphics/devMode/page1/Page1.java
@@ -4,6 +4,10 @@ import java.util.List;
 import java.awt.*;
 import java.awt.event.ActionListener;
 
+import graphics.devMode.page1.header.CommonHeader;
+import graphics.devMode.page1.header.HeaderButtonListener;
+import graphicsEngine.panels.BorderProperties;
+import graphicsEngine.presets.panels.AbstractHeader;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -17,6 +21,8 @@ import graphics.devMode.common.DevPage;
 
 //TODO: add javadocs
 public class Page1 extends DevPage {
+    private HeaderButtonListener headerButtonListener;
+
     private Page1() {
         this(null, null, null);
     }
@@ -37,6 +43,25 @@ public class Page1 extends DevPage {
     //TODO: add javadoc
     public static @NotNull String getStaticPageKey() {
         return (new Page1()).getPageKey();
+    }
+
+    //TODO: add javadoc
+    @Override
+    public boolean addParticularListener(@Nullable ActionListener listener) {
+        if (listener instanceof HeaderButtonListener) {
+            headerButtonListener = (HeaderButtonListener) listener;
+            return true;
+        }
+        return false;
+    }
+
+    /**
+     * TODO: finish this javadoc
+     */
+    @Override
+    public @Nullable AbstractHeader getHeader(@Nullable SimpleColorScheme headerColors,
+                                              @Nullable BorderProperties borderProperties) {
+        return new CommonHeader(headerColors, borderProperties, headerButtonListener);
     }
 
     //TODO: add javadoc
