+++ src/main/java/graphics/devMode/page1/header/Button_Page1.java
@@ -1,4 +1,4 @@
-package graphics.devMode.common.header;
+package graphics.devMode.page1.header;
 
 import java.awt.event.ActionListener;
